% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link-data.R
\name{get_link_data}
\alias{get_link_data}
\title{Create a link data.frame}
\usage{
get_link_data(mantel_df, corr_df, type = c("upper", "lower"),
  var_grp_name = NULL, var_mat_name = NULL, show_diag = FALSE,
  grp_hjust = NULL, grp_vjust = NULL, mat_hjust = NULL,
  mat_vjust = NULL)
}
\arguments{
\item{mantel_df}{mantel's test data.frame, must be returned by \code{\link[ggmantel]{fortify_mantel}}.}

\item{corr_df}{correalation's test data.frame, must be returned by \code{\link[ggcorrr]{fortify_corr}}.}

\item{type}{correalation matrix plot's type.}

\item{var_grp_name}{the column name of link group variable in `mantel_df`.}

\item{show_diag}{logical, if TRUE, correalation matrix plot will don't show diagonal.}

\item{grp_hjust}{a numeric vector to adjust the horizonal position of group points.}

\item{grp_vjust}{a numeric vector to adjust the vertical position of group points.}

\item{mat_hjust}{a numeric vector to adjust the horizonal position of correalation matrixgroup points.}

\item{mat_vjust}{a numeric vector to adjust the vertical position of correalation matrixgroup points.}

\item{var_grp_name}{the column name of link matrix variable in `mantel_df`.}
}
\description{
At present, this function is still in the experimental stage and only
supports the results of \code{\link[ggmantel]{fortify_mantel}} and \code{\link[ggcorrr]{fortify_corr}}.
}
\examples{
library(ggcorrr)
library(ggmantel)
library(vegan)
data(varespec)
data(varechem)
spec <- list(spec01 = varespec[ , 22:25],
           spec02 = varespec[ , 1:4],
            spec03 = varespec[ , 38:43],
            spec04 = varespec[ , 15:20])
df <- fortify_mantel(spec, varechem, process = FALSE)
corr_df <- fortify_corr(varechem, type = "upper", show_diag = FALSE,
                      corr_test = TRUE, cluster = T)
mantel_df <- get_link_data(df, corr_df, type = "upper", grp_hjust = c(0, -1.5, -1, 0))
}
\seealso{
\code{\link[ggmantel]{fortify_mantel}}, \code{\link[ggcorrr]{fortify_corr}}
}
